+++ core/src/main/java/com/microsoft/applicationinsights/internal/agent/CoreAgentNotificationsHandler.java
@@ -27,6 +27,7 @@ import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.Date;
 import java.util.LinkedList;
 
 import com.microsoft.applicationinsights.agent.internal.coresync.AgentNotificationsHandler;
@@ -166,6 +167,32 @@ final class CoreAgentNotificationsHandler implements AgentNotificationsHandler {
     }
 
     @Override
+    public void httpMethodFinishedWithPath(String identifier, String method, String path, String correlationId, String uri, String target, int result, long delta) {
+        if (!LocalStringsUtils.isNullOrEmpty(uri) && (uri.startsWith("https://dc.services.visualstudio.com") || uri.startsWith("https://rt.services.visualstudio.com"))) {
+            return;
+        }
+        long deltaInMS = nanoToMilliseconds(delta);
+        String name = method + " " + path;
+        RemoteDependencyTelemetry telemetry = new RemoteDependencyTelemetry(name, uri, new Duration(deltaInMS), true);
+        Date dependencyStartTime = new Date(System.currentTimeMillis() - deltaInMS);
+        telemetry.setTimestamp(dependencyStartTime);
+        telemetry.setId(correlationId);
+        telemetry.setResultCode(Integer.toString(result));
+        telemetry.setType("HTTP");
+        if (target != null && !target.isEmpty()) {
+            if (telemetry.getTarget() == null) {
+                telemetry.setTarget(target);
+            } else {
+                telemetry.setTarget(telemetry.getTarget() + " | " + target);
+            }
+        }
+
+        InternalLogger.INSTANCE.trace("'%s' sent an HTTP method: '%s', uri: '%s', duration=%s ms", identifier, method, uri, deltaInMS);
+        telemetryClient.track(telemetry);
+
+    }
+
+    @Override
     public void jedisMethodStarted(String name) {
         int index = name.lastIndexOf('#');
         if (index != -1) {            
@@ -213,7 +240,9 @@ final class CoreAgentNotificationsHandler implements AgentNotificationsHandler {
     public void methodFinished(String classAndMethodNames, long deltaInNS, Object[] args, Throwable throwable) {
         long durationInMS = nanoToMilliseconds(deltaInNS);
         Duration duration = new Duration(durationInMS);
+        Date dependencyStartTime = new Date(System.currentTimeMillis() - durationInMS);
         RemoteDependencyTelemetry telemetry = new RemoteDependencyTelemetry(classAndMethodNames, null, duration, throwable == null);
+        telemetry.setTimestamp(dependencyStartTime);
         telemetry.setDependencyKind(DependencyKind.Other);
 
         if (args != null) {
@@ -371,10 +400,11 @@ final class CoreAgentNotificationsHandler implements AgentNotificationsHandler {
     }
 
     private void sendInstrumentationTelemetry(MethodData methodData, Throwable throwable) {
-        Duration duration = new Duration(nanoToMilliseconds(methodData.interval));
-        RemoteDependencyTelemetry telemetry = new RemoteDependencyTelemetry(methodData.name, null, duration, throwable == null);
+        long durationInMilliSeconds = nanoToMilliseconds(methodData.interval);
+        RemoteDependencyTelemetry telemetry = new RemoteDependencyTelemetry(methodData.name, null, new Duration(durationInMilliSeconds), throwable == null);
         telemetry.setType(methodData.type);
-
+        Date dependencyStartDate = new Date(System.currentTimeMillis() - durationInMilliSeconds);
+        telemetry.setTimestamp(dependencyStartDate);
         InternalLogger.INSTANCE.trace("Sending RDD event for '%s'", methodData.name);
 
         telemetryClient.track(telemetry);
@@ -394,6 +424,8 @@ final class CoreAgentNotificationsHandler implements AgentNotificationsHandler {
 
             RemoteDependencyTelemetry telemetry = new RemoteDependencyTelemetry(url, null, duration, throwable == null);
             telemetry.setDependencyKind(DependencyKind.Http);
+            Date dependencyStartDate = new Date(System.currentTimeMillis() - durationInMilliSeconds);
+            telemetry.setTimestamp(dependencyStartDate);
             telemetryClient.trackDependency(telemetry);
             if (throwable != null) {
                 ExceptionTelemetry exceptionTelemetry = new ExceptionTelemetry(throwable);
@@ -430,7 +462,9 @@ final class CoreAgentNotificationsHandler implements AgentNotificationsHandler {
                     duration,
                     throwable == null);
             telemetry.setDependencyKind(DependencyKind.SQL);
-
+            Date dependencyStartTime = new Date(System.currentTimeMillis() - durationInMilliSeconds);
+            telemetry.setTimestamp(dependencyStartTime);
+  
             StringBuilder sb = null;
             if (methodData.arguments.length > 3) {
                 sb = formatAdditionalSqlArguments(methodData);