+++ core/src/main/java/com/microsoft/applicationinsights/internal/channel/common/ActiveTransmissionNetworkOutput.java
@@ -22,9 +22,9 @@
 package com.microsoft.applicationinsights.internal.channel.common;
 
 import java.util.concurrent.RejectedExecutionException;
-import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import com.google.common.base.Preconditions;
 import com.microsoft.applicationinsights.internal.channel.TransmissionOutput;
@@ -38,14 +38,13 @@ public final class ActiveTransmissionNetworkOutput implements TransmissionOutput
     private final static int DEFAULT_MIN_NUMBER_OF_THREADS = 7;
     private final static int DEFAULT_MAX_NUMBER_OF_THREADS = 7;
     private final static long DEFAULT_REMOVE_IDLE_THREAD_TIMEOUT_IN_SECONDS = 60L;
+    private final static AtomicInteger INTSTANCE_ID_POOL = new AtomicInteger(1);
 
     private final int maxThreads;
-
     private final ThreadPoolExecutor outputThreads;
-
     private final TransmissionOutput actualOutput;
-
     private final TransmissionPolicyStateFetcher transmissionPolicy;
+    private final int instanceId = INTSTANCE_ID_POOL.getAndIncrement();
 
     public ActiveTransmissionNetworkOutput(TransmissionOutput actualOutput, TransmissionPolicyStateFetcher transmissionPolicy) {
         this(actualOutput, transmissionPolicy, DEFAULT_MAX_MESSAGES_IN_BUFFER);
@@ -63,14 +62,7 @@ public final class ActiveTransmissionNetworkOutput implements TransmissionOutput
                 maxThreads,
                 DEFAULT_REMOVE_IDLE_THREAD_TIMEOUT_IN_SECONDS,
                 maxMessagesInBuffer);
-        outputThreads.setThreadFactory(new ThreadFactory() {
-            @Override
-            public Thread newThread(Runnable r) {
-                Thread thread = new Thread(r);
-                thread.setDaemon(true);
-                return thread;
-            }
-        });
+        outputThreads.setThreadFactory(ThreadPoolUtils.createDaemonThreadFactory(ActiveTransmissionNetworkOutput.class, instanceId));
     }
 
     @Override