+++ web/src/main/java/com/microsoft/applicationinsights/web/internal/correlation/AppProfileFetcher.java
@@ -21,14 +21,16 @@
 
 package com.microsoft.applicationinsights.web.internal.correlation;
 
+import java.io.Closeable;
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
+
 import org.apache.http.ParseException;
 
 /**
  * Retrieves the application profile from storage
  */
- public interface AppProfileFetcher {
+ public interface AppProfileFetcher extends Closeable {
      /**
       * Fetches the application profile and returns the appId corresponding to the
       * instrumentation key provided.